conductor(checkpoint): Checkpoint end of Phase 1 - Monetary Unit Canonicalization
Canonicalize all monetary values to integer cents across the entire stack:

- Migration: Convert all monetary columns from decimal to bigint
- Models: Cast monetary fields as integer (not float)
- Domain logic: Remove all float-based currency arithmetic
- Frontend: formatCurrency() divides cents by 100 for display
- Factories/seeders: All monetary values in integer cents
- Tests: All 274 tests pass (1459 assertions)

Boundary contract: backend = integer cents, frontend display = dollars

+++ app/Http/Middleware/HandleInertiaRequests.php
@@ -109,7 +109,7 @@ class HandleInertiaRequests extends Middleware
         );
 
         return [
-            'cash' => (float) $gameState->cash,
+            'cash' => (int) $gameState->cash,
             'xp' => $gameState->xp,
             'day' => $gameState->day,
             'level' => $this->calculateLevel($gameState->xp),
